+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 - Export notes with file detail
 - Add unix socket support
 - 設定ファイルでioredisの全てのオプションを指定可能に
+- 通報をDiscordのWebhookに送信できるように
 - Fix: エクスポートしたカスタム絵文字のzipが大きいと読み込めない問題を修正
 - Fix: リモートサーバーに無意味なActivityPubの配信を行うことがあるのを修正
 - Fix: Remove Meilisearch index when notes are deleted